% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/btw_pro.R
\name{btw_pro}
\alias{btw_pro}
\title{A treap function}
\usage{
btw_pro(ppi, pgi, genesym)
}
\arguments{
\item{ppi}{A two-column matrix of PPIs. Each row contains the two nodes of an interaction entry.}

\item{pgi}{A two-column matrix of PGIs. Each row contains the two nodes of an interaction entry.}

\item{genesym}{A vector of all genes in adjusted p-values.}
}
\description{
This function calculates betweenness values of a PPI+PGI network, which is 
obtained from combining protein-protein interactions, or PPIs, and protein-gene 
interactions, or PGIs. Interactions with genes not available in adjusted p-values 
will be removed.
}
\examples{
btw_pro()
}
