% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_pro.R
\name{rank_pro}
\alias{rank_pro}
\title{A treap function}
\usage{
rank_pro(x, direction)
}
\arguments{
\item{x}{A numeric vector.}

\item{direction}{This defines the sorting order. For "large", larger values in x will have higher ranks. For "small", smaller values in x will have higher ranks.}
}
\description{
This function allows customized sorting and returns the ranks of values in a 
vector. Same values will have the same ranks.
}
\examples{
rank_pro()
}
