% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treap.R
\name{treap}
\alias{treap}
\title{A treap function}
\usage{
treap(adj, btw)
}
\arguments{
\item{adj}{A matrix of djusted p-values. The column names refer to samples to be analyzed, and the row names refer to genes.}

\item{btw}{A vector of PPI+PGI betweenness. The names of btw should be the same type of gene IDs with adj, but don't have to be in the same order.}
}
\description{
This function calculates TREAP scores for each gene given adjusted p-values from 
drug treated gene expression profiles and betweenness values from a network of 
protein-protein interactions, or PPIs, and protein-gene interactions, or PGIs. 
Genes with higher scores are more likely to be targets of associated drugs.
}
\examples{
treap()
}
